requirements.txt
Flask==3.0.0
pymongo==4.6.0
python-dotenv==1.0.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
APScheduler==3.10.4
waitress==2.1.2
